Moved client stuff into client folder to separate from server a bit. Made necessary changes in server.js to avoid conflict.
