--- public/manifest.json
@@ -1,13 +1,0 @@
-{
-  "short_name": "React App",
-  "name": "Create React App Sample",
-  "icons": [{
-    "src": "favicon.ico",
-    "sizes": "192x192",
-    "type": "image/png"
-  }],
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
